+++ src_rebuild/Game/C/system.c
@@ -791,17 +791,17 @@ void SetupDrawBufferData(int num_players)
 	{
 		for (j = 0; j < num_players; j++)
 		{
-			u_long* otpt;
+			u_int* otpt;
 			u_char* primpt;
 
 			if (toggle)
 			{
-				otpt = (u_long*)_OT2;
+				otpt = (u_int*)_OT2;
 				primpt = (u_char*)_primTab2; // _primTab1 + PRIMTAB_SIZE
 			}
 			else
 			{
-				otpt = (u_long*)_OT1;
+				otpt = (u_int*)_OT1;
 				primpt = (u_char*)_primTab1;
 			}
