+++ src_rebuild/Game/C/tile.c
@@ -55,10 +55,10 @@ void Tile1x1Lit(MODEL* model)
 		{
 			prims = (POLY_GT4*)plotContext.primptr;
 
-			*(ulong*)&prims->r0 = plotContext.colour;
-			*(ulong*)&prims->r1 = plotContext.colour;
-			*(ulong*)&prims->r2 = plotContext.colour;
-			*(ulong*)&prims->r3 = plotContext.colour;
+			*(uint*)&prims->r0 = plotContext.colour;
+			*(uint*)&prims->r1 = plotContext.colour;
+			*(uint*)&prims->r2 = plotContext.colour;
+			*(uint*)&prims->r3 = plotContext.colour;
 			setPolyGT4(prims);
 
 			// retrieve first three verts
@@ -162,7 +162,7 @@ void Tile1x1(MODEL *model)
 		{
 			prims = (POLY_FT4*)plotContext.primptr;
 
-			*(ulong*)&prims->r0 = plotContext.colour;
+			*(uint*)&prims->r0 = plotContext.colour;
 			setPolyFT4(prims);
 			
 			// retrieve first three verts
@@ -443,7 +443,7 @@ void drawMesh(MVERTEX(*VSP)[5][5], int m, int n, _pct *pc)
 	for (int index = 0; index < numPolys; index++)
 	{
 		setPolyFT4(prim);
-		*(ulong*)&prim->r0 = pc->colour; // FIXME: semiTransparency support
+		*(uint*)&prim->r0 = pc->colour; // FIXME: semiTransparency support
 
 		// test
 		gte_ldv3(&(*VSP)[index][0], &(*VSP)[index][1], &(*VSP)[index][2]);
@@ -539,10 +539,10 @@ void drawMeshLit(MVERTEX(*VSP)[5][5], int m, int n, _pct* pc)
 
 			gte_stsxy(&prim->x3);
 
-			*(ulong*)&prim->r0 = plotContext.colour;
-			*(ulong*)&prim->r1 = plotContext.colour;
-			*(ulong*)&prim->r2 = plotContext.colour;
-			*(ulong*)&prim->r3 = plotContext.colour;
+			*(uint*)&prim->r0 = plotContext.colour;
+			*(uint*)&prim->r1 = plotContext.colour;
+			*(uint*)&prim->r2 = plotContext.colour;
+			*(uint*)&prim->r3 = plotContext.colour;
 
 			SVECTOR tmpPos;
 			gte_ldv0(&(*VSP)[index][0]);
